+++ tools/emitCommentChange/tool.js
@@ -16,14 +16,15 @@
       const lichess = lt.lichess;
       if (!lichess.analysis) return;
       const $ = lt.$;
-      const observer = $('body').observer();
-      observer.clear();
-      observer.on('comment',this.detectNew,{ 
-        subtree: true,
-        childList: true, 
-        attributes: false, 
-        characterData: true
-      });
+      const observer = $('body')
+        .removeObserver('emitCommentChange')
+        .observer('emitCommentChange')
+        .on('comment',this.detectNew,{ 
+          subtree: true,
+          childList: true, 
+          attributes: false, 
+          characterData: true
+        });
     }
 
   }