+++ tools/exportPgn/tool.js
@@ -10,6 +10,14 @@
         defaultValue: true,
         advanced: true,
         hidden: true
+      },
+      {
+        name: 'exportPGNoptions',
+        category: 'analysis',
+        type: 'multiple',
+        possibleValues: ['exportClock', 'exportEval', 'exportTags'],
+        defaultValue: 'exportClock,exportEval,exportTags',
+        advanced: true
       }
     ];
 
@@ -18,13 +26,21 @@
         'options.analysis': 'Analysis',
         'options.exportPGN': 'Export PGN',
         'PGNCopiedToClipboard': 'PGN copied to clipboard',
-        'clipboardDenied': 'Clipboard access denied'
+        'clipboardDenied': 'Clipboard access denied',
+        'options.exportPGNoptions': 'Options for PGN exports',
+        'exportPGNoptions.exportClock': 'Export clock values',
+        'exportPGNoptions.exportEval': 'Export computer evaluation',
+        'exportPGNoptions.exportTags': 'Export PGN tags'
       },
       'ro-RO': {
         'options.analysis': 'Analiz\u0103',
         'options.exportPGN': 'Export\u0103 PGN',
         'PGNCopiedToClipboard': 'PGN copiat \u00een clipboard',
-        'clipboardDenied': 'Acces refuzat la clipboard'
+        'clipboardDenied': 'Acces refuzat la clipboard',
+        'options.exportPGNoptions': 'Op\u0163iuni pentru exporturi PGN',
+        'exportPGNoptions.exportClock': 'Export\u0103 timp pe mut\u0103ri',
+        'exportPGNoptions.exportEval': 'Export\u0103 evalu\u0103ri computer',
+        'exportPGNoptions.exportTags': 'Export\u0103 etichete PGN'
       }
     }
 
@@ -35,6 +51,9 @@
         toPosition: false,
         separateLines: false,
         unicode: false,
+        exportClock: this.options.exportClock,
+        exportEval: this.options.exportEval,
+        exportTags: this.options.exportTags,
         ...options
       };
       const lt = this.lichessTools;
@@ -94,7 +113,7 @@
           const code = shape.brush[0].toUpperCase() + shape.orig + (shape.dest || '');
           group.shapes.push(code);
         }
-        if (node.clock) {
+        if (options.exportClock && node.clock) {
           const group = {
             type: 'clk',
             shapes: [ centisecondsToClock(node.clock) ]
@@ -102,7 +121,7 @@
           groups.push(group);
         }
         const evl = node.ceval || node.eval;
-        if (evl) {
+        if (options.exportEval && evl) {
           const group = {
             type: 'eval',
             shapes: [ evalToString(evl) ]
@@ -251,22 +270,24 @@
         }
         const varNodes = getVarNodes(varNode, options.separateLines);
         const pgns = [];
-        const tags = lt.clone(analysis.study?.data?.chapter?.tags) || [];
-        if (analysis.getOrientation() != 'white') {
+        const tags = (options.exportTags && lt.clone(analysis.study?.data?.chapter?.tags)) || [];
+        if (options.exportTags && analysis.getOrientation() != 'white') {
           addTag(tags, 'StartFlipped', '1');
           addTag(tags, 'Orientation', 'Black');
         }
         if (varNode?.fen && !lt.isStartFen(varNode.fen)) {
           addTag(tags, 'FEN', varNode.fen);
-          addTag(tags, 'SetUp', '1');
+          if (options.exportTags)addTag(tags, 'SetUp', '1');
         }
-        addTag(tags, 'Site', lt.global.location.href, true);
-        const now = new Date().toISOString();
-        addTag(tags, 'UTCDate', now.substr(0, 10).replaceAll('-', '.'), true);
-        addTag(tags, 'UTCTime', now.substr(11, 8), true);
+        if (options.exportTags) {
+          addTag(tags, 'Site', lt.global.location.href, true);
+          const now = new Date().toISOString();
+          addTag(tags, 'UTCDate', now.substr(0, 10).replaceAll('-', '.'), true);
+          addTag(tags, 'UTCTime', now.substr(11, 8), true);
+        }
         const tagString = tags.length ? tags.map(tag => '[' + tag[0] + ' "' + tag[1] + '"]').join('\r\n') + '\r\n' : '';
         for (const node of varNodes) {
-          const pgn = tagString + renderNodesTxt(node, options.fromPosition);
+          const pgn = tagString + renderNodesTxt(node, true);
           pgns.push(pgn);
         }
         const result = pgns.join('\r\n\r\n');
@@ -284,7 +305,15 @@
     async start() {
       const lt = this.lichessTools;
       const value = lt.currentOptions.getValue('exportPGN');
+      const opts = lt.currentOptions.getValue('exportPGNoptions');
       this.logOption('Export PGN', value);
+      this.logOption('Export PGN options', opts);
+      this.options = {
+        exportClock: lt.isOptionSet(opts, 'exportClock'),
+        exportEval: lt.isOptionSet(opts, 'exportEval'),
+        exportTags: lt.isOptionSet(opts, 'exportTags')
+      };
+
       if (value) {
         lt.exportPgn = this.exportPgn;
       } else {